requirements.txt
numpy
matplotlib
mpl_toolkits.mplot3d.art3d